+++ extension/src/ALObject/ALControl.ts
@@ -1,8 +1,8 @@
+import { isNullOrUndefined } from 'util';
 import * as Common from '../Common';
 import { TransUnit } from "../XLIFFDocument";
 import { ALElement } from "./ALElement";
 import { ALObject } from "./ALObject";
-import { ALPagePart } from "./ALPagePart";
 import { ALProperty } from "./ALProperty";
 import { ALXmlComment } from './ALXmlComment';
 import { ALControlType, ALObjectType, ALPropertyType, MultiLanguageType, XliffTokenType } from "./Enums";
@@ -12,7 +12,6 @@ import { XliffIdToken } from "./XliffIdToken";
 export class ALControl extends ALElement {
     type: ALControlType = ALControlType.None;
     _name?: string;
-    private _value?: string;
     xliffTokenType: XliffTokenType = XliffTokenType.InheritFromControl;
     multiLanguageObjects: MultiLanguageObject[] = new Array();
     controls: ALControl[] = new Array();
@@ -20,14 +19,11 @@ export class ALControl extends ALElement {
     xmlComment?: ALXmlComment;
     isALCode: boolean = false;
 
-    constructor(type: ALControlType, name?: string, value?: string) {
+    constructor(type: ALControlType, name?: string) {
         super();
         this.type = type;
         if (name) {
             this.name = name;
-        }
-        if (value) {
-            this.value = value;
         }
     }
 
@@ -46,60 +42,10 @@ export class ALControl extends ALElement {
         this._name = Common.TrimAndRemoveQuotes(name);
     }
 
-    public get value(): string {
-        if (!this._value) {
-            return '';
-        }
-        return this._value;
-    }
-
-    public set value(value: string) {
-        value = value.trim();
-        if (value.toLowerCase().startsWith('rec.')) {
-            value = value.substr(4);
-        }
-        this._value = Common.TrimAndRemoveQuotes(value);
-    }
 
     public get caption(): string {
         let prop = this.multiLanguageObjects.filter(x => x.name === MultiLanguageType[MultiLanguageType.Caption])[0];
-        if (prop) {
-            return prop.text;
-        }
-        let object = this.getObject();
-        if ([ALObjectType.Page, ALObjectType.PageExtension].includes(object.objectType) && [ALControlType.PageField, ALControlType.Part].includes(this.type)) {
-            // Check table for caption
-            let objects = this.getAllObjects();
-
-            if (objects) {
-                if (this.type === ALControlType.Part) {
-                    let part = <unknown>this;
-                    let castedPart = <ALPagePart>part; // Workaround since "(this instanceof ALPagePart)" fails with "TypeError: Class extends value undefined is not a constructor or null"
-                    let relatedObj = castedPart.relatedObject;
-                    if (relatedObj) {
-                        return relatedObj?.caption;
-                    }
-                } else {
-                    let sourceObject;
-                    if (object.objectType === ALObjectType.Page) {
-                        if (object.sourceTable !== '') {
-                            sourceObject = objects.filter(x => (x.objectType === ALObjectType.Table && x.name === object.sourceTable))[0];
-                        }
-                    } else if (object.objectType === ALObjectType.PageExtension) {
-                        if (object.extendedTableId) {
-                            sourceObject = objects.filter(x => x.objectType === ALObjectType.TableExtension && x.extendedObjectId === object.extendedTableId)[0];
-                        }
-                    }
-                    if (sourceObject) {
-                        const allControls = sourceObject.getAllControls();
-                        const fields = allControls.filter(x => x.type === ALControlType.TableField);
-                        let field = fields.filter(x => x.name === this.value)[0];
-                        return field ? field.caption : '';
-                    }
-                }
-            }
-        }
-        return '';
+        return isNullOrUndefined(prop) ? '' : prop.text;
     }
 
     public get toolTip(): string {
@@ -147,6 +93,10 @@ export class ALControl extends ALElement {
         } else {
             return prop.text;
         }
+    }
+
+    public isIdentical(otherControl: ALControl): boolean {
+        return (otherControl.type === this.type && otherControl.name === this.name);
     }
 
     public getObjectType(): ALObjectType {
